Phase II finish: full month names, property photo, drink/grocery links on cover, dates fix
- Trip dates now use full month names: 'December 8 – December 12, 2026' (not 'Dec 8')
- Property cover photo shows in accommodation block on itinerary cover
- Drink Order + Groceries & Breakfast cards with links added to cover page
- Dates no longer appear 3x: eyebrow is clean 'Name · City Concierge Itinerary'
  without dates; tripName no longer rendered anywhere

+++ src/ConciergePanel.jsx
@@ -2812,7 +2812,7 @@ function EditDrawer({ kickoff, onClose, onSave, onSilentUpdate }) {
     const a = kickoff?.arrivalDate, d = kickoff?.departureDate;
     if (a && d) {
       try {
-        const fmt = (s) => new Date(s + "T12:00:00").toLocaleDateString("en-US", { month: "short", day: "numeric" });
+        const fmt = (s) => new Date(s + "T12:00:00").toLocaleDateString("en-US", { month: "long", day: "numeric" });
         const yr  = new Date(d + "T12:00:00").getFullYear();
         return `${fmt(a)} – ${fmt(d)}, ${yr}`;
       } catch {}
@@ -2853,7 +2853,7 @@ function EditDrawer({ kickoff, onClose, onSave, onSilentUpdate }) {
     const a = kickoff?.arrivalDate2, d = kickoff?.departureDate2;
     if (a && d) {
       try {
-        const fmt = (s) => new Date(s + "T12:00:00").toLocaleDateString("en-US", { month: "short", day: "numeric" });
+        const fmt = (s) => new Date(s + "T12:00:00").toLocaleDateString("en-US", { month: "long", day: "numeric" });
         const yr  = new Date(d + "T12:00:00").getFullYear();
         return `${fmt(a)} – ${fmt(d)}, ${yr}`;
       } catch {}
@@ -2865,7 +2865,7 @@ function EditDrawer({ kickoff, onClose, onSave, onSilentUpdate }) {
   useEffect(() => {
     if (!arrivalDate || !departureDate) return;
     try {
-      const fmt = (s) => new Date(s + "T12:00:00").toLocaleDateString("en-US", { month: "short", day: "numeric" });
+      const fmt = (s) => new Date(s + "T12:00:00").toLocaleDateString("en-US", { month: "long", day: "numeric" });
       const yr  = new Date(departureDate + "T12:00:00").getFullYear();
       setTripDates(`${fmt(arrivalDate)} – ${fmt(departureDate)}, ${yr}`);
     } catch {}
@@ -2876,7 +2876,7 @@ function EditDrawer({ kickoff, onClose, onSave, onSilentUpdate }) {
   useEffect(() => {
     if (!arrivalDate2 || !departureDate2) return;
     try {
-      const fmt = (s) => new Date(s + "T12:00:00").toLocaleDateString("en-US", { month: "short", day: "numeric" });
+      const fmt = (s) => new Date(s + "T12:00:00").toLocaleDateString("en-US", { month: "long", day: "numeric" });
       const yr  = new Date(departureDate2 + "T12:00:00").getFullYear();
       setTripDates2(`${fmt(arrivalDate2)} – ${fmt(departureDate2)}, ${yr}`);
     } catch {}
